+++ src/Storages/MergeTree/MutateTask.cpp
@@ -2,12 +2,12 @@
 
 #include <base/logger_useful.h>
 #include <Common/escapeForFileName.h>
+#include <Parsers/queryToString.h>
+#include <Interpreters/SquashingTransform.h>
 #include <Processors/Transforms/TTLTransform.h>
 #include <Processors/Transforms/TTLCalcTransform.h>
-#include <Common/SquashingTransform.h>
 #include <Processors/Transforms/DistinctSortedTransform.h>
 #include <Processors/Transforms/ColumnGathererTransform.h>
-#include <Parsers/queryToString.h>
 #include <Processors/Sources/SourceFromSingleChunk.h>
 #include <Processors/Transforms/ExpressionTransform.h>
 #include <Processors/Transforms/MaterializingTransform.h>
